Fix lints in the Prime app
