+++ src/apps/prime/api/episode-capabilities.ts
@@ -7,6 +7,21 @@ import { ITitleInfo, PrimeApi } from "../api";
 
 const debug = _debug("babbling:prime:episodes");
 
+function mapEpisodes(info: ITitleInfo) {
+    if (info.episodes == null) {
+        return [];
+    }
+
+    return info.episodes
+        .filter(raw => raw.episodeNumber >= 1)
+        .map(raw => ({
+            indexInSeason: raw.episodeNumber - 1,
+            season: raw.seasonNumber - 1,
+            title: raw.title,
+            titleId: raw.titleId,
+        }));
+}
+
 export interface IPrimeEpisode extends IEpisodeBase {
     titleId: string;
 }
@@ -47,15 +62,4 @@ export class PrimeEpisodeCapabilities implements IEpisodeCapabilities<IPrimeEpis
             yield mapEpisodes(season);
         }
     }
-}
-
-function mapEpisodes(info: ITitleInfo) {
-    return info.episodes!
-        .filter(raw => raw.episodeNumber >= 1)
-        .map((raw, index) => ({
-            indexInSeason: raw.episodeNumber - 1,
-            season: raw.seasonNumber - 1,
-            title: raw.title,
-            titleId: raw.titleId,
-        }));
 }
