+++ src/apps/prime/channel.ts
@@ -13,13 +13,66 @@ import { EpisodeResolver } from "../../util/episode-resolver";
 
 // NOTE: this sure looks like a circular dependency, but we're just
 // importing it for the type definition
-import { IPrimeOpts, PrimeApp } from ".";
+import type { IPrimeOpts, PrimeApp } from ".";
 
 import { PrimeApi } from "./api";
 import { PrimeEpisodeCapabilities } from "./api/episode-capabilities";
 import { AvailabilityType, IAvailability, ISearchResult } from "./model";
 
 const debug = _debug("babbling:PrimeApp:player");
+
+function isAvailableOnlyWithAds(availability: IAvailability[]) {
+    const canPlayWithAds = availability.findIndex(
+        a => a.type === AvailabilityType.FREE_WITH_ADS,
+    ) !== -1;
+    if (!canPlayWithAds) return false;
+
+    // we can play with ads, so it's *only* available with ads iff we don't find
+    // another availability type
+    return availability.findIndex(a => a.type === AvailabilityType.PRIME
+            || a.type === AvailabilityType.OTHER_SUBSCRIPTION
+            || a.type === AvailabilityType.OWNED) === -1;
+}
+
+function pickTitleIdFromUrl(url: string) {
+    const m1 = url.match(/video\/detail\/([^/]+)/);
+    if (m1) {
+        return m1[1];
+    }
+
+    const m2 = url.match(/gp\/product\/([^/?]+)/);
+    if (m2) {
+        return m2[1];
+    }
+
+    const m3 = url.match(/dp\/([^/?]+)/);
+    if (m3) {
+        return m3[1];
+    }
+}
+
+function playableForMovieById(id: string) {
+    return async (app: PrimeApp, opts: IPlayableOptions) => {
+        if (opts.resume === false) {
+            await app.play(id, { startTime: 0 });
+        } else {
+            await app.play(id, {});
+        }
+    };
+}
+
+function playableFromTitleId(titleId: string) {
+    return async (app: PrimeApp, _opts: IPlayableOptions) => app.resumeSeriesByTitleId(titleId);
+}
+
+function playableFromSearchResult(result: ISearchResult) {
+    if (result.type === ContentType.MOVIE) {
+        // we can use MOVIE results directly
+        return playableForMovieById(result.titleId || result.id);
+    }
+
+    return playableFromTitleId(result.titleId);
+}
 
 interface IPrimeResultExtras {
     titleId: string;
@@ -128,55 +181,4 @@ export class PrimePlayerChannel implements IPlayerChannel<PrimeApp> {
             };
         }
     }
-}
-
-function isAvailableOnlyWithAds(availability: IAvailability[]) {
-    const canPlayWithAds = availability.findIndex(a => a.type === AvailabilityType.FREE_WITH_ADS) !== -1;
-    if (!canPlayWithAds) return false;
-
-    // we can play with ads, so it's *only* available with ads iff we don't find
-    // another availability type
-    return availability.findIndex(a => a.type === AvailabilityType.PRIME
-            || a.type === AvailabilityType.OTHER_SUBSCRIPTION
-            || a.type === AvailabilityType.OWNED) === -1;
-}
-
-function pickTitleIdFromUrl(url: string) {
-    const m1 = url.match(/video\/detail\/([^\/]+)/);
-    if (m1) {
-        return m1[1];
-    }
-
-    const m2 = url.match(/gp\/product\/([^\/\?]+)/);
-    if (m2) {
-        return m2[1];
-    }
-
-    const m3 = url.match(/dp\/([^\/\?]+)/);
-    if (m3) {
-        return m3[1];
-    }
-}
-
-function playableFromSearchResult(result: ISearchResult) {
-    if (result.type === ContentType.MOVIE) {
-        // we can use MOVIE results directly
-        return playableForMovieById(result.titleId || result.id);
-    }
-
-    return playableFromTitleId(result.titleId);
-}
-
-function playableFromTitleId(titleId: string) {
-    return async (app: PrimeApp, opts: IPlayableOptions) => app.resumeSeriesByTitleId(titleId);
-}
-
-function playableForMovieById(id: string) {
-    return async (app: PrimeApp, opts: IPlayableOptions) => {
-        if (opts.resume === false) {
-            await app.play(id, { startTime: 0 });
-        } else {
-            await app.play(id, {});
-        }
-    };
 }
